Split global.tpl.h into types.h and global.tpl.h
